requirements.txt
appnope==0.1.0
ask-sdk-core==1.11.0
ask-sdk-model==1.15.0
ask-sdk-runtime==1.11.0
ask-sdk-webservice-support==1.0.0
asn1crypto==1.1.0
astroid==2.3.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
flask-ask-sdk==1.0.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mstrio-py==10.11.1
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.2
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pylint==2.4.2
pyparsing==2.4.3
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
